requirements.txt
fastapi
uvicorn
httpx
presidio-analyzer
spacy
fpdf2
